all non get requests require authentication

+++ api/controllers/areaController.js
@@ -15,24 +15,54 @@ const readArea = (req, res) => {
 };
 
 const addArea = (req, res) => {
+    const isAdmin = req.isAdmin;
+
+    if (!isAdmin) {
+        return res.status(403).send('Not authorized.');
+    }
+
     handlePromise(areaActions.addArea(req.body), res);
 };
 
 const editArea = (req, res) => {
+    const isAdmin = req.isAdmin;
+
+    if (!isAdmin) {
+        return res.status(403).send('Not authorized.');
+    }
+
     const id = req.params.id;
     handlePromise(areaActions.editArea(id, req.body), res);
 };
 
 const deleteArea = (req, res) => {
+    const isAdmin = req.isAdmin;
+
+    if (!isAdmin) {
+        return res.status(403).send('Not authorized.');
+    }
+
     const id = req.params.id;
     handlePromise(areaActions.deleteArea(id), res);
 };
 
 const truncateAreas = (req, res) => {
+    const isAdmin = req.isAdmin;
+
+    if (!isAdmin) {
+        return res.status(403).send('Not authorized.');
+    }
+
     handlePromise(areaActions.truncateAreas(), res);
 };
 
 const addAreas = (req, res) => {
+    const isAdmin = req.isAdmin;
+
+    if (!isAdmin) {
+        return res.status(403).send('Not authorized.');
+    }
+    
     handlePromise(areaActions.addAreas(req.body), res);
 };
